# portals/common/core/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "apk-site Admin / Release / Ops API",
    "version": "2.0.0",
    "description": "Auto-generated from Flask url_map (admin portal mode)."
  },
  "paths": {
...
    "/api/ops-platform/runtime/active": {
      "get": {
        "summary": "project ops.ops platform runtime active",
        "operationId": "project_ops.ops_platform_runtime_active_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/ops-platform/runtime/flow-control": {
      "post": {
        "summary": "project ops.ops platform runtime flow control",
        "operationId": "project_ops.ops_platform_runtime_flow_control_post",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/ops-platform/runtime/flow-status": {
      "get": {
        "summary": "project ops.ops platform runtime flow status",
        "operationId": "project_ops.ops_platform_runtime_flow_status_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/status": {
      "get": {
        "summary": "api.status",
        "operationId": "api.status_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}